fixing ageSex api tests
