+++ server/api/ageSex.js
@@ -5,14 +5,30 @@ module.exports = router
 router.get('/', async (req, res, next) => {
   try {
     console.log('in try')
-    const nation = await AgeSex.findAll({
+    const nationData = await AgeSex.findAll({
       where: {
         state: 'United States'
       }
     })
     console.log('after db find')
-    console.log(nation)
-    res.json(nation)
+    console.log(nationData)
+    res.json(nationData)
+  } catch (error) {
+    next(error)
+  }
+})
+
+router.get('/:stateName', async (req, res, next) => {
+  try {
+    console.log('in try')
+    const stateData = await AgeSex.findAll({
+      where: {
+        state: req.params.stateName
+      }
+    })
+    console.log('after db find')
+    console.log(stateData)
+    res.json(stateData)
   } catch (error) {
     next(error)
   }
