+++ server/db/db.js
@@ -3,19 +3,19 @@ const pkg = require('../../package.json')
 
 const databaseName = pkg.name + (process.env.NODE_ENV === 'test' ? '-test' : '')
 
-const db = new Sequelize(
-  process.env.DATABASE_URL || `postgres://localhost:5432/${databaseName}`,
-  {
-    logging: false
-  }
-)
-
 // const db = new Sequelize(
-//   process.env.DATABASE_URL || `postgres://ania:newPassword@localhost/${databaseName}`,
+//   process.env.DATABASE_URL || `postgres://localhost:5432/${databaseName}`,
 //   {
 //     logging: false
 //   }
 // )
+
+const db = new Sequelize(
+  process.env.DATABASE_URL || `postgres://ania:newPassword@localhost/${databaseName}`,
+  {
+    logging: false
+  }
+)
 
 module.exports = db
